requirements.txt
telethon==0.11.5
db-sqlite3==0.0.1
psycopg2-binary==3.0.0
